+++ featurebyte/schema/use_case.py
@@ -1,9 +1,9 @@
 """
 Use Case API payload schema
 """
-from typing import List, Optional
+from typing import Any, Dict, List, Optional
 
-from pydantic import Field, StrictStr
+from pydantic import Field, StrictStr, root_validator
 
 from featurebyte.models.base import FeatureByteBaseModel, PydanticObjectId
 from featurebyte.models.use_case import UseCaseModel
@@ -21,6 +21,15 @@ class UseCaseCreate(FeatureByteBaseModel):
     target_namespace_id: Optional[PydanticObjectId]
     context_id: PydanticObjectId
     description: Optional[StrictStr]
+
+    @root_validator(pre=True)
+    @classmethod
+    def _validate_target(cls, values: Dict[str, Any]) -> Dict[str, Any]:
+        target_id = values.get("target_id", None)
+        target_namespace_id = values.get("target_namespace_id", None)
+        if not target_id and not target_namespace_id:
+            raise ValueError("Either target_id or target_namespace_id must be specified.")
+        return values
 
 
 class UseCaseUpdate(BaseDocumentServiceUpdateSchema):
